+++ usr/tgtd.h
@@ -57,7 +57,7 @@ extern int target_cmd_queue(int host_no, uint8_t *scb, unsigned long uaddr,
 			    uint8_t *lun, uint32_t data_len,
 			    int attribute, uint64_t tag);
 extern void target_cmd_done(int host_no, uint64_t tag);
-extern void target_mgmt_request(int host_no, int req_id, int function,
+extern void target_mgmt_request(int host_no, uint64_t req_id, int function,
 				uint8_t *lun, uint64_t tag);
 
 extern void target_cmd_io_done(void *key, int result);
